Del: JsonString, JsonMap. Add: planet.json package -> JsonString, JsonMapList.
